Refactor magic numbers, DRY code, and improve modularity in fd-vscode
- Modularize `extension.ts`:
  - Move `FdDiagnosticsProvider` to `src/diagnostics.ts`
  - Move `FdTreePreviewPanel` to `src/panels/tree-preview.ts`
  - Move `FdSpecViewPanel` to `src/panels/spec-view.ts`
  - Move `FdDocumentSymbolProvider` to `src/document-symbol.ts`
  - Extract HTML template and constants to `src/webview-html.ts`
- Refactor `ai-refine.ts`:
  - Extract constants for configuration and default models
  - DRY up API call logic into `callAiApi`
- Refactor `fd-parse.ts`:
  - Extract regex patterns into constants
  - Improve readability
- Refactor `e2e-ux.test.ts`:
  - Use constants for shapes and colors
- Ensure all tests pass with `npm test`
